Further simplify processing of public IPs
